+++ frontend/src/components/chat-input-loading.tsx
@@ -25,6 +25,7 @@ export default function ChatInputLoading({ newMessage }: ChatComponentProps) {
             name="prompt"
             placeholder="Message the AI..."
             autoComplete="off"
+            disabled
           />
           <div className="flex gap-2 mr-3">
             <div className="w-6 h-6 border-2 ml-2 p-1 border-gray-500 border-t-transparent rounded-full animate-spin" />